ci(docker): use existing application setting
The application already has a PREREVIEW_ORCID_SANDBOX setting; the value in the YAML file was incorrect as it needs to
be a string rather than a boolean, so this removes it.

I'm not sure if this is the cause of the dev environment failing; I'd be surprised, but without feedback it's impossible
to tell.

Refs #390