test(integration): make sure the homepage looks correct
Adds a visual-regression test for the homepage using Playwright: this makes sure that it looks the same in Chromium on
a large screen.

It blocks requests to Twitter to stop the timeline in the footer from changing, and resets both the carousels to the
first item.

It uses Git LFS to store the reference screenshots, since they might be quite large.

Refs #388