feat: let Fathom analytics read from the environment
Rather than having to build the site ID into the image, this allows it to be read at runtime from
the environment. This involves having to pass it from the backend to the frontend, as only the
former has access.

There is a JavaScript error in development (which doesn't use the backend for serving the
frontend), and so the magic __FATHOM_SITEID__ string isn't replaced.

Refs #252