test(integration): hide rather than block the Twitter timeline
Using page.route() prevents browser caching from working, which will help speed up the integration
tests (once resources are allowed to be cached). This instead hides the Twitter timeline from the
screenshot. This has a slight difference in that the space it occupies is present, hence the
difference in the iPhone homepage reference image.

This does mean that the Twitter timeline needs a consistent space, whether it's loaded yet or not.
This is done by moving the height from the timeline to its parent element.

I'm not sure why the requesting-a-review reference image also changes; I've not been able to
recreate the 2-line button effect locally. It does now look more like the existing iPhone image
though.

Refs #388