perf(caching): enable HTTP caching for static assets
Allows browsers to cache (and reuse) assets that can't change. This sets a conservative limit of
1 minute, which is enough for reuse when clicking around the site but not for repeated visits. This
is to give a chance for problems to be found and fixed, as clearing browser caches is hard at best.
A more realistic max-age will be 1 year (the maximum value that is safe to use).

This recognises files named like 'index.337f472b.js', that is they have a hash of their content in
their name. As a result, a change of content results in a new file, leading to a fresh download.

Refs #392