build(lfs): let the staging workflow checkout Git LFS assets
The staging workflow is failing with an unhelpful error message. I think the problem might be that
assets are now stored in Git LFS rather than the Git repo directly, and so it will need to check
them out separately.

This does also show that the staging workflow should not be rebuilding the image.

Refs #392, 0cbb848