ci: separate build and deploy jobs
This is a start at breaking up the monolithic workflows, by separating the build and deploy jobs and running them in sequence. This means that each job can be focused, rather than having to be aware of the rest of the process. For example, only the deploy job needs to know about the Azure CLI.

Refs #395