ci: fix image tag configuration
The GitHub Actions syntax in a384f78 was incorrect, as the IMAGE_TAG variable was not interpolated and didn't exist in the Azure environment.

The Docker commands were working without interpolation, however, as the IMAGE_TAG is available in their environment.

As the deploy job passed when it should not have, I've opened https://github.com/Azure/webapps-deploy/issues/189.

Refs #395