ci: inline registry username
GitHub Actions hides any secrets from the logs; the registry username is stored as a secret. As the username is just the name of the app itself the logs are hard to read, since unrelated uses of the word are hidden.

Usernames aren't secrets, they can be stored in the environment.

Refs #395