test(integration): add case for finding and reviewing a request
Adds a test case for searching for a preprint that has a review request and viewing it.

The search page changes slightly when there is data, causing other reference images to be updated.

The API allows for a user to request a review more than once (which I'm not sure is expected), so this tries to make sure it's only done once. Ideally, the fixture would be set up only when it's needed.

Refs #388