build(docker): mount an empty folder for the integration test results
GitHub Actions has seen some failures due to permission problems. This is caused by the mounted folder not existing, so Docker has to create it. On Linux, the permissions are based on the container, which causes a failure when the host tries to subsequently write to it. (This doesn't happen on Macs due to the Docker For Mac intermediary virtual machine.)

So, this makes sure the folder exists but is empty before mounting.

Refs #388