fix: only use the content URL from Crossref if the content type is set
Turns out the failures I've been seeing in CI are the result having triggered a bug locally: it should not be possible to see a HTML page embedded, only a PDF. The metadata resolver was finding a content URL from Crossref (HTML) but using the content type from Google Scholar (PDF). This tries to group the content URL and type in the Crossref response together, so it uses both of them together or neither.

Refs #388