build(docker): require commands in the entry point script to pass
By default, a shell script carries on even if a step fails. If, for example, a migration fails, the app starts up anyway with an unknown and possibly dangerous result.

This change causes a failure in any command to fail the whole script, except in cases where it's acceptable (as we test the exit code).

Refs #390, #395