build: generate code without building
Reduce the steps in running restful-react by generating the OpenAPI doc and running the converter as part of the build separately. As a result, the script doesn't have to be built by Parcel.

This change also ensures that the file is built and shared with the host correctly when running the development environment.

Refs #399