fix: allow logging in when using production mode locally
It hasn't been possible to log in when running production mode locally. I've just discovered that it's the use of NODE_ENV causing broken configuration when not running the site on a prereview.org domain. 66f2dfd had caused the NODE_ENV to be 'production' locally.

Using the NODE_ENV to change settings will always create problems; instead, we can calculate the setting based on the ORCID callback URL, which will work no matter where we run the site.

Refs #395