test(integration): add a case for requesting a review of a preprint
Adds a test case that sees a logged-in user find a preprint and request a review.

As each test needs to request the first review of a preprint, a unique preprint is generated for the combination of test name and environment. As a result, the separate Chrome and iPhone runs don't collide, but they still see the same unique preprint each time.

Unfortunately, this does mean that changing the test title will cause a different preprint to be used (and so the reference images will need updating). Still, I can't think of another identification mechanism.

Refs #388