test(integration): add a case for submitting a rapid review
Adds a test case that sees a logged-in user find a preprint and submit a rapid review.

This also adds a case that checks that the form has to be completed before submitting.

I've had to increase the timeout for tests as the first is a little long due to the number of form fields to check. I tried running them in parallel, but Playwright threw an error.

Refs #388