perf(images): don't load banners on the Communities page
Images inlined in the API are one of the leading performance problems. The community banner images can be heavy (one of the three is 1.8MB), which is the leading cause of the slowdown on the listing page.

This commit prevents the banner images from being loaded for the listing page. As they're decorative rather than part of the core content, it's far safer to remove them than enter the rabbit hole of trying to fix them.

Refs #392, 440b3e3