perf: don't show community filters when preprints are part of a list
When preprints are part of a list, data for all the communities are loaded to show a review filter. Unfortunately, rather than requesting just the names for communities that it needs to display, it loads all the data for all communities (including the list of all the communities members, which is 599 for one of them). The filters don't share the information, so the browser has to make a separate API request for each preprint shown.

All review lists already have a by-community filter, and others are already filtered somehow (e.g. for an individual community). So, this filter seems redundant.

This change removes the filter, except for on the preprint page (where it is loaded only once).

Refs #392