perf: split code on routes
While not a single-page application, the site has been rendered in the browser using a single, bloated JavaScript file. Introducing caching has helped, but any change to the code can still cause the file to changed and so re-downloaded.

This change splits the file into smaller chunks that are loaded when needed. While this doesn't address the fundamental problem of rendering in the browser, it should improve the situation a bit:

1. Less JavaScript is loaded, so it'll be quicker and use device resources (memory, data plan etc).
2. Changes to the code have less of an impact, as only parts have to be re-downloaded. Again, this makes it quicker and uses fewer resources.

Refs #392