test(integration): regenerate iPhone 11 homepage reference image
The regenerated image in 2548130 for the homepage didn't include the Twitter block at the bottom of the page. The fact that this happened suggests there's a race condition, but having tried quite a few times, I couldn't recreate it.

CI runs with retries enabled, so that should prevent it from happening there.

Refs 1588af7