fix(permissions): allow community moderators to edit templates
Community moderators have the option to edit a community template, but the API requires admin access. This change opens it to anyone with permission to edit the community.

There's a new endpoint for editing a community's template rather than changing the permissions for the existing endpoint to edit a template. Ideally, we wouldn't need a new endpoint, but the permissions logic seems complicated and given there are no tests, I don't want to touch it. Copying and pasting the endpoint logic from the existing PUT and DELETE controllers has fewer risks.

Refs #410