test(integration): partially add a case for using a template
Adds a test case for a community member using a template when publishing a full review.

Unfortunately, as the functionality only copies the template to the clipboard rather than loading it into the editor, Playwright can't yet complete the test case.

I tried to follow the example in https://github.com/microsoft/playwright/issues/2511#issuecomment-641617914, but it didn't work in either Chromium or Safari. In the former, I think it wasn't recognised as a secure origin (as the hostname is 'prereview', not 'localhost'). In the latter, Safari did not recognise the 'clipboard-read' permission at all.

Refs #388, #410