ci: check that the migration files are up to date
When the code is changed, we might need migration files to ensure the database has the structure that the code expects. The ORM library can generate these, but they have to be manually run and checked. This change adds a check to CI to ensure that they are updated (i.e. the ORM doesn't find any differences).

I expect CI to fail, as they are not currently up to date.

Refs #388, #395, #400