test(integration): generate consistent dates
A test started to fail as it used a reference image containing a generated date, which is only valid for runs on that date. Until today, runs must not have produced enough pixel differences to fail.

This change uses libfaketime to change the system date so that generated dates are always the same. I'm using the 'start at' option ('@') to see time differences in the logs (it's always midnight otherwise). I'm not aware of anywhere in the app that displays times, so this shouldn't be a problem.

The date I've chosen is the one that appears in the failing test, solely to avoid regenerating those images.

Refs ccde93a