ci: only store integration test results if the step ran
Use of failure() in GitHub Actions causes a step to run if any previous step fails. We want only to try and store the integration test results if that particular step fails; otherwise, there's nothing to do.

We still have to use the global failure() condition as steps don't run if something fails.

Refs #388, #395