test(jest) allow tests to run for longer
I can't work out why Jest is timing out in CI; I'm not able to reproduce it locally. To see if it really is just being slow, this increases the timeout for a test to be 30 seconds rather than 5.

It also disables ts-jest's diagnostics, which should improve performance slightly too.

Refs ba79ea2, #388