build: pin Alpine version
The default Alpine version has changed from 3.11 to 3.14, which seems to have a different setup leading to the Sharp build not working (which seems similar to the failures when trying to upgrade Node to version 16). This change pins the version to the highest supported version that still works without any other adjustments.

Refs https://github.com/nodejs/docker-node/pull/1596, https://github.com/PREreview/prereview/issues/421#issuecomment-948550009