ci(git-lfs): cache Git LFS files
GitHub's Git LFS offering has both bandwidth and storage limits, and @dasaderi received an email saying we've reached the former. Git LFS in GitHub Actions is still subject to the bandwidth limit (even though it's all inside GitHub's network).

This change introduces a cache for the LFS objects to reduce the impact of CI: it should now only use the Git LFS bandwidth when one of the objects changes (e.g. adding a new reference image in the integration tests).

Refs https://github.com/actions/checkout/issues/165#issuecomment-657673315