test(integration): don't try and recreate preprints
If a test fails, retries attempt to recreate the same preprint. This change adds a check to whether it needs to create it.

The main problem is that the integration tests all run against the same app, sharing state. There's more than can be done here, but this seems enough for the moment.

Refs #388, 89646b7