test(integration): normalise the page when taking a locator screenshot
The page normalisation steps only run when taking a screenshot of the entire page or the current screen. This change allows them to run when taking a screenshot of a locator too.

The most important of these steps is the removal of transitions, as there have been race conditions in the Material UI click effect.

Refs #388, bb4d72d