test(integration): hide all CSS animations and transitions
We've continued to see race conditions on the Material UI ripple effect; it turns out that it's a CSS animation, so they need to be disabled too.

I've also updated the animation and transition normalisation to make it run extremely fast rather than not at all. This change means that JavaScript listening for an animation event would still work.

Refs #388, dfeffbe, https://github.com/mui-org/material-ui/issues/16367#issuecomment-569829175