test(frontend): add cases for the rapid reviews bar chart
This change adds test cases for the bar chart that shows for rapid reviews on a preprint. It includes a failing case too.

As the first frontend test, it uses JSDOM (https://github.com/jsdom/jsdom) and Testing Library (https://testing-library.com/). Beyond the failing test case, it also shows some design problems that make the component hard to test.

Refs #388, #409