fix: don't hide script failures
Currently, if a script (e.g. database migrations) fails, the error is shown to the console and then swallowed. The app will then be in an unknown state, leading to strange behaviour.

This change makes sure that the script throws the error and the resulting exit code halts the process.

Refs #395