fix: show add review/request button to community members
This change returns community information for a user's default persona, which allows the button to be displayed. Without it, the user appears as not being a member of any community, so the app would never show the button.

I'm not sure if there's any other logic that depends on this information, so there's a chance it might fix/create bugs in untested code.

Fixes #429