fix: prevent multiple requests
This change checks to see if a user has already created a request by one of their personas and returns an error if so.

It doesn't do this at the database level, so it is vulnerable to a race condition. Doing so would involve changing the schema to allow for a user- rather than persona-based index and removing existing duplicates. This more straightforward change should be acceptable for now, though.

Fixes #432