fix: reduce the size of a user response
A user cannot log into the site as the request to the user API is failing. Node runs out of memory when serializing the JSON response due to many duplicate/recursive entries compounded by embedded images. I believe that the response is trying to include details of a single community 560 times, including its 2.1MB banner image.

This change crudely cuts out some of the duplication/recursive behaviour and some of the images. Given that the browser makes this request on every page, this change could have unwanted side effects.