fix: disable Google Scholar search
This change disables Google Scholar as a source of preprint metadata. It searched for the DOI, scraped the HTML results, picked the first one, then filled in gaps left by the Crossref API and scraping the DOI itself. However, checking that the first result is the one we're after is impossible (it doesn't provide a DOI to compare against). The integration tests broke as Google Scholar started returning a different order for the test DOI '10.5555/12345678': an actual arXiv article that referenced it currently appears first. As a result, the metadata is a mixture of DOI 10.5555/12345678 and arXiv article 1809.01194 (leading to failing tests). I don't know if this has happened in actual usage, but as preprint metadata is never updated, it seems dangerous to continue this. It's unclear how much value searching Google Scholar adds.